+++ flake.nix
@@ -23,7 +23,7 @@
 
       buildOpnix = pkgs.buildGoModule {
         pname = "opnix";
-        version = "0.9.0";
+        version = "0.10.0";
         inherit src;
         inherit vendorHash;
         subPackages = ["cmd/opnix"];