+++ nix/checks.nix
@@ -6,7 +6,7 @@
   # Run tests
   go-tests = pkgs.buildGoModule {
     pname = "opnix-go-tests";
-    version = "0.9.0";
+    version = "0.10.0";
     inherit src;
     inherit vendorHash;
 
@@ -20,7 +20,7 @@
   # Run golangci-lint
   go-lint = pkgs.buildGoModule {
     pname = "opnix-go-lint";
-    version = "0.9.0";
+    version = "0.10.0";
     inherit src;
     inherit vendorHash;